third_party/blink/web_tests/platform/ios/external/wpt/css/cssom-view/range-bounding-client-rect-with-display-contents-expected.txt

This is a testharness.js-based test.
[FAIL] the space between elements using a range should be the same as using another method
  assert_equals: range.getBoundingClientRect().height expected 47 but got 60
Harness: the test ran to completion.
